change weight from int to double

+++ Cardealership/src/cardealership/cars/Car.java
@@ -61,7 +61,7 @@ public abstract class Car {
         return maxSpeed.getSpeed();
     }
     
-    public int performWeight(){
+    public double performWeight(){
         return weight.getWeight();
     }
